check auth and Clearing Post-User Relations

+++ controllers/feed.js
@@ -1,5 +1,6 @@
 const Post = require("../model/post");
 const removeFile = require("../middlewares/removeFile");
+const User = require("../model/user");
 
 exports.getPosts = async (req, res, next) => {
   try {
@@ -53,15 +54,17 @@ exports.createPost = async (req, res, next) => {
     title,
     content,
     imageUrl,
-    creator: {
-      name: "ahmed",
-    },
+    creator: req.userId,
   });
   try {
     const result = await newPost.save();
+    const user = await User.findById(req.userId);
+    user.posts.push(result);
+    const updateUser = await user.save();
     res.status(201).json({
       message: "post created successfully!",
       post: result,
+      creator: { _id: user._id, name: user.name },
     });
   } catch (err) {
     if (!err.statusCode) {
@@ -87,6 +90,12 @@ exports.updatePost = async (req, res, next) => {
 
   try {
     const post = await Post.findOne({ _id: postId });
+    const postCreatorId = post.creator.toString() || "";
+    if (postCreatorId !== req.userId) {
+      const err = new Error("not authorized!");
+      err.statusCode = 422;
+      return next(err);
+    }
     if (title === post.title) {
       title = undefined;
     }
@@ -101,7 +110,6 @@ exports.updatePost = async (req, res, next) => {
     if (!(title || content || imageUrl)) {
       const err = new Error("no changes");
       err.statusCode = 422;
-      /*     throw err; */
       return next(err);
     }
     const result = await Post.findByIdAndUpdate(
@@ -134,13 +142,27 @@ exports.updatePost = async (req, res, next) => {
 exports.deletePost = async (req, res, next) => {
   const postId = req.params.postId;
   try {
-    const result = await Post.findByIdAndDelete(postId);
-    if (!result) {
+    const post = await Post.findById(postId);
+    if (!post) {
       const err = new Error("post not found!");
       err.statusCode = 422;
-      /*     throw err; */
       return next(err);
     }
+    const user = await User.findById(req.userId);
+    const postCreatorId = post.creator.toString() || "";
+    if (postCreatorId !== req.userId) {
+      const err = new Error("not authorized!");
+      err.statusCode = 422;
+      return next(err);
+    }
+    const result = await Post.findByIdAndDelete(postId);
+    const updatedPosts = user.posts.filter((p) => {
+      return p.toString() !== postId;
+    });
+    user.posts = updatedPosts;
+    // this is another method in mongoose
+    //  user.posts.pull(postId);
+    const updateUser = await user.save();
     removeFile(undefined, result.imageUrl);
     res.status(200).json({
       message: "post deleted successfully!",
